pyproject.toml
[project]
name = "kindl-notebook-importer"
version = "0.1.0"
description = "The Project converts Kindle Notebook entries to Makrdown-Files"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "google-generativeai>=0.8.4",
    "pyperclip>=1.9.0",
]
